// code/SDP-webfront/SDP/src/components/color-cascader/package/types.ts
import { IDictionaryItem } from '@/hooks/use-dictionary/types';

export type IColorItem = Omit<IDictionaryItem, 'children'> & {
  /** 父子编码用-拼接 */
  pathCode: string;
  /** 颜色英文缩写 */
  colorAbbrCode: string;
  /** 色号 */
  colorNumber: string;
  /** 颜色英文名 */
  colorEnglishName: string;
  coverUrl?: string;
  children: IColorItem[];
};